fronius_scraper: Factor most scraper funs into module
  Allows for easier definition of scripts for hatch.
